Move utils to utility
